[bessel-refactor] Match recurrence loop counter type to CPU
Math.cuh's bessel_k_recurrence used 'int n' for the loop counter while
Math.h uses 'int64_t n'. N is bounded above by ~2000 today (the
nu > 2000 dispatch goes to bessel_k_uniform_asymptotic), so this is
not a functional issue, but the asymmetry is needless and would
silently break if the dispatch threshold ever changed.

No behavior change for current dispatch.

Authored with assistance from Claude.

+++ aten/src/ATen/native/cuda/Math.cuh
@@ -3048,7 +3048,7 @@ const auto modified_bessel_k_string = modified_bessel_i_string + modified_bessel
         T K_curr = K_mu1;
         T log_scale = T(0.0);
 
-        for (int n = 1; n < N; n++) {
+        for (int64_t n = 1; n < N; n++) {
             T order = mu + T(n);
             T K_next = K_prev + (T(2.0) * order / x) * K_curr;
             K_prev = K_curr;
